mysite/requirement.txt
aiohttp==3.3.2
apilimiter==1.0.4
APScheduler==3.5.3
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
backports.functools-lru-cache==1.5
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
cheroot==6.5.2
click==6.7
cryptography==2.4.1
cssselect==1.0.3
Django==2.1.3
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
get==1.0.3
idna==2.7
idna-ssl==1.1.0
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.4
Markdown==3.0.1
MarkupSafe==1.0
more-itertools==4.3.0
multidict==4.3.1
Pillow==5.3.0
post==1.0.2
public==2.0.1
pycparser==2.19
Pygments==2.3.0
PyMySQL==0.9.2
pyquery==1.4.0
pytz==2018.7
query-string==1.0.2
redis==2.10.6
request==1.0.2
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.14
tzlocal==1.5.1
urllib3==1.23
uWSGI==2.0.17.1
virtualenv==16.1.0
web.py==0.40.dev1
Werkzeug==0.14.1
yarl==1.2.6
